feat(camel-config): sentinel data-node auth
- redis repo backends accept password/username/db (0..=16383) in
  sentinel mode: data-node credentials were inexpressible before,
  NOAUTH under requirepass; db selection was always 0
- camel-redis: RedisEndpointConfig gains username, db widens u8->u16
  (both URI parsers cap 16383), set_username propagation in
  sentinel_node_conn_info
- prefix-collision identity includes the effective sentinel db
- redacting Debug for the new credential fields (positive *** tests)
- live: authenticated sentinel topology, NOAUTH negative control
- closes bd rc-vkpl; follow-ups rc-q3mb rc-eztw rc-swzq rc-1xc8

Bd: rc-vkpl

+++ crates/components/camel-redis/src/metadata.rs
@@ -31,7 +31,7 @@ pub(super) struct RedisMetadataDescriptor {
     pub _password: String,
 
     #[uri_param(name = "db", default = "0")]
-    pub _db: u8,
+    pub _db: u16,
 
     #[uri_param(name = "ssl")]
     pub _ssl: Option<bool>,
